Team_Workspace/Bae_JH/d_0324/Main_Docker_Runtime/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
langchain
langchain-core
langchain-openai
langchain-google-genai
openai
python-dotenv
python-multipart
psycopg2-binary
GeoAlchemy2
shapely
unstructured
PyMuPDF
aiohttp